requirements.txt
agate==1.10.2
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
annotated-types==0.6.0
anyio==4.3.0
Babel==2.15.0
certifi==2024.2.2
csvkit==2.0.0
dbfread==2.0.7
distro==1.9.0
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
leather==0.4.0
mo-dots==9.578.24081
mo-future==7.546.24057
mo-imports==7.546.24057
mo-parsing==8.581.24094
mo-sql-parsing==10.627.24135
numpy==1.26.4
olefile==0.47
openai==1.30.1
openpyxl==3.1.2
pandas==2.2.2
parsedatetime==2.6
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
text-unidecode==1.3
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
xlrd==2.0.1
